feat: add white noise generator
- Add WhiteNoiseConfigScreen for user to config prompt, duration and prompt influence
- Add WhiteNoiseScreen for show progress and retry
- update string resources for sound effect
- update Elevenlabs to support prompt influence as float

+++ elevenlabs/src/commonMain/kotlin/io/github/kkoshin/elevenlabs/api/SoundGeneration.kt
@@ -13,7 +13,7 @@ class SoundGeneration
 suspend fun ElevenLabsClient.makeSoundEffects(
     prompt: String,
     durationSeconds: Double?,
-    promptInfluence: Double,
+    promptInfluence: Float,
     outputStream: OutputStream
 ): Result<Unit> =
     post<SoundGenerationRequest, SoundGeneration, ByteReadChannel>(
